Model/requirements.txt
pandas>=1.5
numpy>=1.23
scipy>=1.9
scikit-learn>=1.1
networkx>=3.0
python-igraph>=0.10
leidenalg>=0.9
python-louvain>=0.16
umap-learn>=0.5
matplotlib>=3.5
seaborn>=0.12
statsmodels>=0.13
joblib>=1.2
openpyxl>=3.1
tensorflow>=2.10
ompute>=0.0